+++ superblockify/population/__init__.py
@@ -1,0 +1,1 @@
+"""Population init, subpackage for the GHSL Population data"""